requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
deepgram-sdk==5.3.2
toml==0.10.2
python-multipart==0.0.12
websockets==13.1
